requirements.txt
Flask
mss
Pillow
